@@ -3,12 +3,14 @@
 	"version": "1.1.0",
 	"description": "Plugin that extracts metadata and saves it in JSON format",
 	"main": "src/main.ts",
+	"type": "module",
 	"scripts": {
 		"dev": "cross-env BUILD=dev node esbuild.mjs",
 		"build": "cross-env BUILD=production node esbuild.mjs",
 		"release": "standard-version",
 		"test": "jest",
-		"format": "npx prettier --write src/"
+		"format": "npx prettier --write src/",
+		"lint": "eslint ."
 	},
 	"standard-version": {
 		"t": "",
@@ -35,21 +37,23 @@
 	"author": "kometenstaub",
 	"license": "MIT",
 	"devDependencies": {
-		"@types/eslint": "^8.4.10",
-		"@types/jest": "^29.2.4",
-		"@types/node": "^18.11.17",
-		"@typescript-eslint/eslint-plugin": "^5.47.0",
-		"@typescript-eslint/parser": "^5.47.0",
-		"cross-env": "^7.0.3",
-		"esbuild": "0.16.10",
+		"@types/eslint": "^9.6.1",
+		"@types/jest": "^30.0.0",
+		"@types/node": "^24.10.1",
+		"@typescript-eslint/eslint-plugin": "^8.48.1",
+		"@typescript-eslint/parser": "^8.48.1",
+		"cross-env": "^10.1.0",
+		"esbuild": "0.27.1",
 		"esbuild-plugin-inline-worker": "^0.1.1",
-		"eslint": "^8.30.0",
-		"jest": "^29.3.1",
-		"obsidian": "^1.1.1",
-		"prettier": "2.8.1",
+		"eslint": "^9.39.1",
+		"eslint-config-prettier": "^10.1.8",
+		"globals": "^16.5.0",
+		"jest": "^30.2.0",
+		"obsidian": "^1.10.3",
+		"prettier": "3.7.4",
 		"standard-version": "^9.5.0",
-		"ts-jest": "^29.0.3",
-		"tslib": "^2.4.1",
-		"typescript": "^4.9.4"
+		"ts-jest": "^29.4.6",
+		"tslib": "^2.8.1",
+		"typescript": "^5.9.3"
 	}
 }
